Struct rsp2_newtype_indices::Indexed

pub struct Indexed<I: Idx, V: ?Sized> {
    pub raw: V,
    // some fields omitted
}

A Vec or slice that uses newtype indices.

V is only ever [T] or Vec<T>.

Fields
raw: V
Implementations

impl<I: Idx, T> Indexed<I, Vec<T, Global>>

Construction

pub fn from_raw(raw: Vec<T>) -> Self

pub fn new() -> Self

pub fn from_elem<S>(elem: T, universe: &Indexed<I, [S]>) -> Self where
T: Clone,

pub fn from_elem_n(elem: T, n: usize) -> Self where
T: Clone,

pub fn with_capacity(capacity: usize) -> Self

impl<I: Idx, T> Indexed<I, [T]>

Construction

pub fn from_raw_ref(raw: &[T]) -> &Self

pub fn from_raw_mut(raw: &mut [T]) -> &mut Self

impl<I: Idx, T> Indexed<I, Vec<T>>

Length-changing operations

pub fn push(&mut self, d: T) -> I

pub fn pop(&mut self) -> Option<T>

pub fn resize(&mut self, new_len: usize, value: T) where
T: Clone,

pub fn shrink_to_fit(&mut self)

pub fn truncate(&mut self, a: usize)

impl<I: Idx, T> Indexed<I, [T]>

Simple access

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn last_index(&self) -> Option<I>

pub fn binary_search(&self, value: &T) -> Result<I, I> where
T: Ord,

pub fn get(&self, index: I) -> Option<&T>

pub fn get_mut(&mut self, index: I) -> Option<&mut T>

pub fn pick2_mut(&mut self, a: I, b: I) -> (&mut T, &mut T)

Return mutable references to two distinct elements, a and b. Panics if a == b.

impl<I: Idx, T> Indexed<I, Vec<T>>

By-value iteration

pub fn into_iter(self) -> IntoIter<T>

pub fn into_iter_enumerated(
self
) -> Map<Enumerate<IntoIter<T>>, impl FnMut((usize, T)) -> (I, T)>

Iterate with indices.

Panics

It is guaranteed that calling next on this this iterator will not panic.

impl<I: Idx, T> Indexed<I, [T]>

Non-consuming iteration

pub fn iter(&self) -> Iter<'_, T>

pub fn iter_enumerated<'a>(
&'a self
) -> Map<Enumerate<Iter<'a, T>>, impl FnMut((usize, &'a T)) -> (I, &'a T)>

Iterate with indices.

Panics

It is guaranteed that calling next on this this iterator will not panic.

pub fn indices(&self) -> Map<Range<usize>, impl FnMut(usize) -> I>

Iterate over the indices.

Panics

It is guaranteed that calling next on this this iterator will not panic.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

pub fn iter_enumerated_mut<'a>(
&'a mut self
) -> Map<Enumerate<IterMut<'a, T>>, impl FnMut((usize, &'a mut T)) -> (I, &'a mut T)>

Iterate with indices.

Panics

It is guaranteed that calling next on this this iterator will not panic.

impl<I: Idx, T> Indexed<I, Vec<T>>

Conversion

pub fn into_index_type<Ix: Idx>(self) -> Indexed<Ix, Vec<T>>

impl<I: Idx, T> Indexed<I, [T]>

Conversion

pub fn as_index_type<Ix: Idx>(&self) -> &Indexed<Ix, [T]>

pub fn as_index_type_mut<Ix: Idx>(&mut self) -> &mut Indexed<Ix, [T]>