Trait rsp2_newtype_indices::IndexCast

pub unsafe trait IndexCast<Result: ?Sized, Disambig: ?Sized> {
    fn index_cast(self) -> Result
    where
        Self: Sized,
        Result: Sized,
    { ... }
}

Trait for index_cast. See the free function for details.

Safety

This is unsafe because it has generic impls that perform transmute.

Special consideration must be given to types whose data layout is affected by trait impls, such as HashMap and BTreeMap.

The following conditions must be met for an impl to be safe:

Self and Result must have identical representations.

If Self implements any of the traits PartialEq, Eq, Hash, PartialOrd, and Ord, then so must Result with an identical implementation (and vice versa).

This implies that generic impls of these traits on implementors of IndexCast must not be marked default.

Notice that this trait is not limited to types for which a safe conversion from Self to Result exists, as it can, for instance, convert &A to &B.

Provided methods

fn index_cast(self) -> Result where
Self: Sized,
Result: Sized,

Loading content...